requirements.txt
# TEAPS v5 - Requirements

# Database
sqlalchemy>=2.0.0
pymysql>=1.1.0

# Telegram Bot
python-telegram-bot>=20.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0

# QR Code & Images

# Authentication & Security
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Environment Variables
python-dotenv>=1.0.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development (optional)
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0
